+++ UA1/examen/dev_15/ua1ex1_teoria.txt
@@ -1,0 +1,33 @@
+NOMBRE:Alejandro Luz Morales 
+FECHA: 16/10/2024
+
+PARTE TEÓRICA
+
+a) Definición de proceso. 1 punto. 
+Un programa tiene procesos, estos son ordenes que da el programa para que se ejecuten correctamente, tambien consumen recursos del sistema (CPU, RAM, Disco Duro...), 
+si hay varios procesos que estan ejecutandose anteriormente que esten consumiendo todos nuestros recursos y añadimos uno nuevo el proceso no se ejecutara ya que no 
+tiene recursos donde hacerlo.
+
+b) Utilizando tus propias palabras, indica la relación entre hilos y procesos. 3 puntos.
+Los hilos son como "mini procesos" que todos juntos forman un proceso, asi que varios hilos pueden estar trabajando a la misma vez en el mismo proceso, si los procesos
+se quieren comunicar entre ellos necesitan al Sistema Operativo, en cambio los hilos pueden comunicarse ellos solos, ademas los hilos comparten recursos lo que hace que 
+resuelvan los probemas mas rapidamente.
+
+c) Utilizando tus propias palabras, indica las características de la programación concurrente y sus ámbitos de aplicación. 3 puntos.
+La programación recurrente se encarga de que los procesos se sincronicen para ejecutarlos a la misma vez. Un ejemplo de uso puede ser: una vodeojuego donde se tienen 
+que ejecutar varios procesos a la vez para que funcionen correctamente.
+
+d) Utilizando tus propias palabras, indica las diferencias entre programación paralela y programación distribuida, sus ventajas e inconvenientes. 3 puntos.
+La programacion paralela es para equipos que tienen varios procesadorees que realizen procesos al mismo tiempo y la programacion distribuida consiste en dividir los 
+componentes en diferentes maquinas.
+
+Las ventajas de programacion paralela son que los procesos se ejecutan a la misma vez todos y se recuden el tiempo de las tareas
+Y los inconvenientes son que es mucho mas caro y necesitas tener varios entornos que a veces pueden dar fallos
+
+Las ventajas de la programacion distribuida es que permite compartir recursos y es mas economico
+Y los inconvenientes son que es mas conplejo el sistema el dividir el hardware y que puede haber fallos de comunicación entre maquinas. 
+
+
+Notas:
+
+El nombre del fichero .txt a entregar debe ser:	examen\dev_X\ua1ex1_teoria.txt , es decir, el fichero ua1ex1_teoria.txt debe estar ubicado en tu carpeta examen\dev_X\
